+++ packages/codegen/src/generate-docs-descriptors.ts
@@ -1,11 +1,10 @@
-import { V14, V15 } from "@polkadot-api/substrate-bindings"
-
 import {
   getChecksumBuilder,
   getLookupFn,
 } from "@polkadot-api/metadata-builders"
 import { getDocsTypesBuilder } from "@/types-builder"
 import { knownTypes } from "./known-types"
+import { UnifiedMetadata } from "@polkadot-api/substrate-bindings"
 
 export type FileTree = {
   [key: string]: string | FileTree
@@ -18,7 +17,7 @@ const paths = {
 
 export async function generateDocsDescriptors(
   key: string,
-  metadata: V14 | V15,
+  metadata: UnifiedMetadata,
 ): Promise<FileTree> {
   const lookup = getLookupFn(metadata)
   const checksumBuilder = getChecksumBuilder(lookup)
@@ -43,6 +42,7 @@ export async function generateDocsDescriptors(
     "EventsFromPalletsDef",
     "ErrorsFromPalletsDef",
     "ConstFromPalletsDef",
+    "ViewFnsFromPalletsDef",
     "SS58String",
     "ResultPayload",
     "TxCallData",
@@ -80,6 +80,11 @@ export async function generateDocsDescriptors(
     docsTypesBuilder,
     getClientImports,
   )
+  const viewFnsOutput = await buildViewFns(
+    metadata,
+    docsTypesBuilder,
+    getClientImports,
+  )
 
   const descriptorsTypesFileContent =
     `import {\n  ${getClientImports().join(",\n  ")}\n} from "${paths.client}";\n` +
@@ -98,7 +103,16 @@ export async function generateDocsDescriptors(
 export type __Circular = any;
 `
 
-  const index = getIndexFileDocs({ chainName: key })
+  const hasSection = {
+    storage: storageOutput.index !== "",
+    runtimeCalls: runtimeCallsOutput.index !== "",
+    errors: errorsOutput.index !== "",
+    constants: constantsOutput.index !== "",
+    events: eventsOutput.index !== "",
+    calls: callsOutput.index !== "",
+    viewFns: viewFnsOutput.index !== "",
+  }
+  const index = getIndexFileDocs({ chainName: key, hasSection })
 
   return {
     index,
@@ -109,11 +123,12 @@ export type __Circular = any;
     Constants: constantsOutput,
     Events: eventsOutput,
     Transactions: callsOutput,
+    ViewFunctions: viewFnsOutput,
   }
 }
 
 const buildEnumObj = <T>(
-  metadata: V14 | V15,
+  metadata: UnifiedMetadata,
   val: number | undefined,
   cb: (name: string, docs: string[]) => T,
 ): Record<string, T> => {
@@ -129,7 +144,7 @@ const buildEnumObj = <T>(
 }
 
 async function buildErrors(
-  metadata: V14 | V15,
+  metadata: UnifiedMetadata,
   docsTypesBuilder: ReturnType<typeof getDocsTypesBuilder>,
   getClientImports: () => string[],
 ): Promise<FileTree> {
@@ -139,7 +154,7 @@ async function buildErrors(
         pallet.name,
         {
           docs: pallet.docs,
-          values: buildEnumObj(metadata, pallet.errors, (name, docs) => {
+          values: buildEnumObj(metadata, pallet.errors?.type, (name, docs) => {
             return {
               type: `PlainDescriptor<${docsTypesBuilder.buildError(
                 pallet.name,
@@ -158,7 +173,7 @@ async function buildErrors(
 }
 
 async function buildConstants(
-  metadata: V14 | V15,
+  metadata: UnifiedMetadata,
   docsTypesBuilder: ReturnType<typeof getDocsTypesBuilder>,
   getClientImports: () => string[],
 ): Promise<FileTree> {
@@ -193,7 +208,7 @@ async function buildConstants(
 }
 
 async function buildEvents(
-  metadata: V14 | V15,
+  metadata: UnifiedMetadata,
   docsTypesBuilder: ReturnType<typeof getDocsTypesBuilder>,
   getClientImports: () => string[],
 ): Promise<FileTree> {
@@ -203,7 +218,7 @@ async function buildEvents(
         pallet.name,
         {
           docs: pallet.docs,
-          values: buildEnumObj(metadata, pallet.events, (name, docs) => {
+          values: buildEnumObj(metadata, pallet.events?.type, (name, docs) => {
             return {
               type: `PlainDescriptor<${docsTypesBuilder.buildEvent(
                 pallet.name,
@@ -222,7 +237,7 @@ async function buildEvents(
 }
 
 async function buildCalls(
-  metadata: V14 | V15,
+  metadata: UnifiedMetadata,
   docsTypesBuilder: ReturnType<typeof getDocsTypesBuilder>,
   getClientImports: () => string[],
 ): Promise<FileTree> {
@@ -232,7 +247,7 @@ async function buildCalls(
         pallet.name,
         {
           docs: pallet.docs,
-          values: buildEnumObj(metadata, pallet.calls, (name, docs) => {
+          values: buildEnumObj(metadata, pallet.calls?.type, (name, docs) => {
             return {
               type: `TxDescriptor<${docsTypesBuilder.buildCall(
                 pallet.name,
@@ -251,7 +266,7 @@ async function buildCalls(
 }
 
 async function buildRuntimeCalls(
-  metadata: V14 | V15,
+  metadata: UnifiedMetadata,
   docsTypesBuilder: ReturnType<typeof getDocsTypesBuilder>,
   getClientImports: () => string[],
 ): Promise<FileTree> {
@@ -282,8 +297,40 @@ async function buildRuntimeCalls(
   return buildTypeFolder(runtimeCalls, getClientImports)
 }
 
+async function buildViewFns(
+  metadata: UnifiedMetadata,
+  docsTypesBuilder: ReturnType<typeof getDocsTypesBuilder>,
+  getClientImports: () => string[],
+): Promise<FileTree> {
+  const viewFns = Object.fromEntries(
+    metadata.pallets.map((pallet) => [
+      pallet.name,
+      {
+        docs: pallet.docs,
+        values: Object.fromEntries(
+          pallet.viewFns.map((fn) => {
+            const { args, value } = docsTypesBuilder.buildViewFn(
+              pallet.name,
+              fn.name,
+            )
+            return [
+              fn.name,
+              {
+                type: `RuntimeDescriptor<${args}, ${value}>`,
+                docs: fn.docs,
+              },
+            ]
+          }),
+        ),
+        descriptorsTypesImports: docsTypesBuilder.recordTypeFileImports(),
+      },
+    ]),
+  )
+  return buildTypeFolder(viewFns, getClientImports)
+}
+
 async function buildStorage(
-  metadata: V14 | V15,
+  metadata: UnifiedMetadata,
   docsTypesBuilder: ReturnType<typeof getDocsTypesBuilder>,
   getClientImports: () => string[],
 ): Promise<FileTree> {
@@ -398,7 +445,21 @@ ${docs.map((doc: string) => ` * ${doc.trim()}`).join("\n")}
 `
 }
 
-function getIndexFileDocs({ chainName }: { chainName: string }): string {
+function getIndexFileDocs({
+  chainName,
+  hasSection,
+}: {
+  chainName: string
+  hasSection: {
+    storage: boolean
+    runtimeCalls: boolean
+    errors: boolean
+    constants: boolean
+    events: boolean
+    calls: boolean
+    viewFns: boolean
+  }
+}): string {
   return `
 /**
  * This is generated documentation for TypedAPI decriptors for **${chainName}** chain  
@@ -422,7 +483,9 @@ function getIndexFileDocs({ chainName }: { chainName: string }): string {
  * @packageDocumentation
  */
 
-/**
+${
+  hasSection.storage
+    ? `/**
  * Storage queries reference
  * 
  * Each item described here is a
@@ -455,7 +518,12 @@ function getIndexFileDocs({ chainName }: { chainName: string }): string {
  */
 export * as Storage from "./Storage";
 
-/**
+`
+    : ""
+}
+${
+  hasSection.constants
+    ? `/**
  * Constants reference
  * 
  * Each item described here is a \`PlainDescriptor<T>\`  
@@ -486,7 +554,12 @@ export * as Storage from "./Storage";
  */
 export * as Constants from "./Constants";
 
-/**
+`
+    : ""
+}
+${
+  hasSection.errors
+    ? `/**
  * Errors
  * 
  * This section is temporarily commented out, 
@@ -498,7 +571,12 @@ export * as Constants from "./Constants";
  */
 // export * as Errors from "./Errors";
 
-/**
+`
+    : ""
+}
+${
+  hasSection.calls
+    ? `/**
  * Transactions reference
  * 
  * Each item described here is a \`TxDescriptor<T>\`, where \`T\` describes
@@ -527,7 +605,12 @@ export * as Constants from "./Constants";
  */
 export * as Transactions from "./Transactions";
 
-/**
+`
+    : ""
+}
+${
+  hasSection.events
+    ? `/**
  * Events
  * 
  * Each item described here is a \`PlainDescriptor<T>\`  
@@ -552,7 +635,40 @@ export * as Transactions from "./Transactions";
  */
 export * as Events from "./Events";
 
-/**
+`
+    : ""
+}
+${
+  hasSection.viewFns
+    ? `/**
+ * View Functions
+ * 
+ * Each item described here is a \`RuntimeDescriptor<Args, ReturnType>\`
+ * 
+ * For example, \`Proxy.is_superset\` is of type
+ * \`\`\`ts
+ * is_superset: RuntimeDescriptor<[to_check: ProxyType, against: ProxyType], boolean>
+ * \`\`\`
+ * and can be called like this:
+ * \`\`\`ts
+ *  const isSuperset = await api.view.Proxy.is_superset(Enum("Any"), Enum("NonTransfer"))
+ *  console.log(isSuperset)
+ * })
+ * \`\`\`
+ * 
+ * @see [PAPI docs](https://papi.how/typed/view) on view functions for more
+ * 
+ * @namespace
+ * @category TypedApi calls
+ */
+export * as ViewFunctions from "./ViewFunctions";
+
+`
+    : ""
+}
+${
+  hasSection.runtimeCalls
+    ? `/**
  * Runtime calls
  * 
  * Each item described here is a \`RuntimeDescriptor<Args, ReturnType>\`
@@ -575,6 +691,9 @@ export * as Events from "./Events";
  */
 export * as RuntimeCalls from "./RuntimeCalls";
 
+`
+    : ""
+}
 /**
  * Descriptors types
  * 
@@ -604,6 +723,5 @@ export * as RuntimeCalls from "./RuntimeCalls";
  * @namespace
  * @category Types
  */
-export * as Types from "./types";
-  `
+export * as Types from "./types";`
 }