+++ packages/tx-utils/src/get-metadata.ts
@@ -1,12 +1,13 @@
-import { V15, V14, decAnyMetadata } from "@polkadot-api/substrate-bindings"
+import {
+  decAnyMetadata,
+  UnifiedMetadata,
+  unifyMetadata,
+} from "@polkadot-api/substrate-bindings"
 import { HexString } from "@polkadot-api/substrate-bindings"
 
-export const getMetadata = (input: Uint8Array | HexString): V14 | V15 => {
+export const getMetadata = (input: Uint8Array | HexString): UnifiedMetadata => {
   try {
-    const { metadata } = decAnyMetadata(input)
-    if (metadata.tag !== "v14" && metadata.tag !== "v15")
-      throw new Error("Wrong metadata version")
-    return metadata.value
+    return unifyMetadata(decAnyMetadata(input))
   } catch (e) {
     throw e || new Error("Unable to decode metadata")
   }