+++ packages/client/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+### Added
+
+- Add `view` API.
+- Expose `<key>Apis` type for Runtime APIs.
+- **Signer:**
+  - `withMetadataHash` enhancer
+
+### Changed
+
+- Prefer metadata V16 when creating the client.
+
 ### Fixed
 
 - **CLI:**