+++ packages/codegen/CHANGELOG.md
@@ -2,9 +2,14 @@
 
 ## Unreleased
 
+### Added
+
+- Add `viewFns` code generation
+- Expose `<key>Apis` type for Runtime APIs.
+
 ### Fixed
 
-- fix entry docs not getting stripped properly [#1033](https://github.com/polkadot-api/polkadot-api/issues/1033).
+- Fix entry docs not getting stripped properly [#1033](https://github.com/polkadot-api/polkadot-api/issues/1033).
 
 ## 0.14.0 - 2025-04-24