+++ packages/signers/ledger-signer/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Add support to `ecdsa` (ETH-like) accounts.
+
 ## 0.1.12 - 2025-04-24
 
 ### Fixed